+++ simson/plastics/plastics_export.py
@@ -1,7 +1,7 @@
 import flodym as fd
 from typing import TYPE_CHECKING
 
-from simson.common.custom_export import CommonDataExporter
+from simson.common.common_export import CommonDataExporter
 
 if TYPE_CHECKING:
     from simson.plastics.plastics_model import PlasticsModel
