+++ simson/steel/steel_export.py
@@ -4,7 +4,7 @@ import flodym as fd
 from typing import TYPE_CHECKING
 import flodym.export as fde
 
-from simson.common.custom_export import CommonDataExporter
+from simson.common.common_export import CommonDataExporter
 from simson.common.common_cfg import SteelVisualizationCfg
 
 if TYPE_CHECKING:
